+++ lib/agens/supervisor.ex
@@ -14,17 +14,8 @@ defmodule Agens.Supervisor do
     strategy: :one_for_one
   )
   ```
-
-  ### Options
-    * `:prefixes` (`Agens.Prefixes`) - The default prompt prefixes can be overriden with this option. Each `Agens.Serving.Config` can also override the defaults on a per-serving basis.
-
-  See the [README.md](README.md#configuration) for more info.
   """
   use Supervisor
-
-  alias Agens.Prefixes
-
-  @default_opts [prefixes: Prefixes.default()]
 
   @doc false
   @spec child_spec(keyword()) :: Supervisor.child_spec()
@@ -38,25 +29,22 @@ defmodule Agens.Supervisor do
 
   @doc false
   @spec start_link(keyword()) :: Supervisor.on_start()
-  def start_link(args) do
-    override = Keyword.get(args, :opts, [])
-    opts = Keyword.merge(@default_opts, override)
-
-    Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
+  def start_link(_args) do
+    Supervisor.start_link(__MODULE__, :ok, name: __MODULE__)
   end
 
   @doc false
   @impl true
-  @spec init(keyword()) ::
+  @spec init(:ok) ::
           {:ok,
            {:supervisor.sup_flags(),
             [:supervisor.child_spec() | (old_erlang_child_spec :: :supervisor.child_spec())]}}
           | :ignore
-  def init(opts) do
+  def init(:ok) do
     children = [
       {Registry, keys: :unique, name: Agens.Registry},
       {Task.Supervisor, name: Agens.JobSupervisor},
-      {Agens, name: Agens, opts: opts}
+      {Agens, name: Agens}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)