+++ test/support/backend.ex
@@ -88,7 +88,7 @@ defmodule Test.Support.Backend do
     nodes = %{
       "sub_node_0" => %Agens.Job.Node{
         serving: :test_serving,
-        agent_id: :sub_final_agent
+        agent_id: "sub_final_agent"
       }
     }